Address non-blocking SVS review items: move IT to :sandbox, Lombok, since 3.8.0, pin custom SVS URL
Land non-design review items separately so the extension-point
discussion (#3370) stays focused on the Query/JNIService/KNNEngine layers.

- Move FaissSVSVamanaIT from core src/test into the :sandbox module and give
  :sandbox its own integTest RestIntegTestTask (testFixtures(project(':')) +
  test cluster + bundlePlugin); drop the default-build exclusions in build.gradle.
  Also disables dependencyLicenses/thirdPartyAudit/loggerUsageCheck for the
  sandbox module, which the new testclusters/RestIntegTestTask plumbing requires.
- Use Lombok @Log4j2 and @NoArgsConstructor on SandboxEngine.
- Bump the experimental 'since' from 3.7.0 to 3.8.0 across the SVS sandbox files.
- Require -DSVS_RUNTIME_SHA256 for a custom -DSVS_RUNTIME_URL (dev-only bypass
  via -DSVS_RUNTIME_ALLOW_UNVERIFIED=ON).
