+++ sandbox/src/main/java/org/opensearch/knn/sandbox/svs/FaissSVSVamanaSearchContext.java
@@ -19,7 +19,7 @@ import static org.opensearch.knn.sandbox.svs.SVSConstants.METHOD_PARAMETER_SEARC
  * recall/latency knob, analogous to {@code ef_search}); {@code search_buffer_capacity} is an index-time
  * parameter only.
  */
-@ExperimentalAlgorithm(description = "Intel SVS Vamana search context", since = "3.7.0")
+@ExperimentalAlgorithm(description = "Intel SVS Vamana search context", since = "3.8.0")
 public final class FaissSVSVamanaSearchContext implements KNNLibrarySearchContext {
 
     @Override
